problem1: 기능 별로 파일 분리
